--- a/MonoGame.Framework/Color.cs
+++ b/MonoGame.Framework/Color.cs
@@ -162,7 +162,10 @@ namespace Microsoft.Xna.Framework
             Yellow = new Color(0xff00ffff);
             YellowGreen = new Color(0xff32cd9a);
         }
-	// ARGB
+
+        // Stored as RGBA with R in the least significant octet:
+        // |-------|-------|-------|-------
+        // A       B       G       R
         private uint _packedValue;
 	  
         /// <summary>
@@ -174,10 +177,6 @@ namespace Microsoft.Xna.Framework
         public Color(uint packedValue)
         {
             _packedValue = packedValue;
-			// ARGB
-			//_packedValue = (packedValue << 8) | ((packedValue & 0xff000000) >> 24);
-			// ABGR			
-			//_packedValue = (packedValue & 0xff00ff00) | ((packedValue & 0x000000ff) << 16) | ((packedValue & 0x00ff0000) >> 16);
         }
 
         /// <summary>
@@ -185,13 +184,8 @@ namespace Microsoft.Xna.Framework
         /// </summary>
         /// <param name="color">A <see cref="Vector4"/> representing color.</param>
         public Color(Vector4 color)
+            : this((int)(color.X * 255), (int)(color.Y * 255), (int)(color.Z * 255), (int)(color.W * 255))
         {
-            _packedValue = 0;
-			
-			R = (byte)MathHelper.Clamp(color.X * 255, Byte.MinValue, Byte.MaxValue);
-            G = (byte)MathHelper.Clamp(color.Y * 255, Byte.MinValue, Byte.MaxValue);
-            B = (byte)MathHelper.Clamp(color.Z * 255, Byte.MinValue, Byte.MaxValue);
-            A = (byte)MathHelper.Clamp(color.W * 255, Byte.MinValue, Byte.MaxValue);
         }
 
         /// <summary>
@@ -199,13 +193,8 @@ namespace Microsoft.Xna.Framework
         /// </summary>
         /// <param name="color">A <see cref="Vector3"/> representing color.</param>
         public Color(Vector3 color)
+            : this((int)(color.X * 255), (int)(color.Y * 255), (int)(color.Z * 255))
         {
-            _packedValue = 0;
-
-            R = (byte)MathHelper.Clamp(color.X * 255, Byte.MinValue, Byte.MaxValue);
-            G = (byte)MathHelper.Clamp(color.Y * 255, Byte.MinValue, Byte.MaxValue);
-            B = (byte)MathHelper.Clamp(color.Z * 255, Byte.MinValue, Byte.MaxValue);
-            A = 255;
         }
 
         /// <summary>
@@ -215,12 +204,16 @@ namespace Microsoft.Xna.Framework
         /// <param name="alpha">The alpha component value from 0 to 255.</param>
         public Color(Color color, int alpha)
         {
-            _packedValue = 0;
+            if ((alpha & 0xFFFFFF00) != 0)
+            {
+                var clampedA = (uint)MathHelper.Clamp(alpha, Byte.MinValue, Byte.MaxValue);
 
-            R = color.R;
-            G = color.G;
-            B = color.B;
-            A = (byte)MathHelper.Clamp(alpha, Byte.MinValue, Byte.MaxValue);
+                _packedValue = (color._packedValue & 0x00FFFFFF) | (clampedA << 24);
+            }
+            else
+            {
+                _packedValue = (color._packedValue & 0x00FFFFFF) | ((uint)alpha << 24);
+            }
         }
 
         /// <summary>
@@ -228,14 +221,9 @@ namespace Microsoft.Xna.Framework
         /// </summary>
         /// <param name="color">A <see cref="Color"/> for RGB values of new <see cref="Color"/> instance.</param>
         /// <param name="alpha">Alpha component value from 0.0f to 1.0f.</param>
-        public Color(Color color, float alpha)
+        public Color(Color color, float alpha):
+            this(color, (int)(alpha * 255))
         {
-            _packedValue = 0;
-
-            R = color.R;
-            G = color.G;
-            B = color.B;
-            A = (byte)MathHelper.Clamp(alpha * 255, Byte.MinValue, Byte.MaxValue);
         }
 
         /// <summary>
@@ -245,13 +233,20 @@ namespace Microsoft.Xna.Framework
         /// <param name="g">Green component value from 0.0f to 1.0f.</param>
         /// <param name="b">Blue component value from 0.0f to 1.0f.</param>
         public Color(float r, float g, float b)
+            : this((int)(r * 255), (int)(g * 255), (int)(b * 255))
         {
-            _packedValue = 0;
-			
-            R = (byte)MathHelper.Clamp(r * 255, Byte.MinValue, Byte.MaxValue);
-            G = (byte)MathHelper.Clamp(g * 255, Byte.MinValue, Byte.MaxValue);
-            B = (byte)MathHelper.Clamp(b * 255, Byte.MinValue, Byte.MaxValue);
-            A = 255;
+        }
+
+        /// <summary>
+        /// Constructs an RGBA color from scalars representing red, green, blue and alpha values.
+        /// </summary>
+        /// <param name="r">Red component value from 0.0f to 1.0f.</param>
+        /// <param name="g">Green component value from 0.0f to 1.0f.</param>
+        /// <param name="b">Blue component value from 0.0f to 1.0f.</param>
+        /// <param name="alpha">Alpha component value from 0.0f to 1.0f.</param>
+        public Color(float r, float g, float b, float alpha)
+            : this((int)(r * 255), (int)(g * 255), (int)(b * 255), (int)(alpha * 255))
+        {
         }
 
         /// <summary>
@@ -262,20 +257,20 @@ namespace Microsoft.Xna.Framework
         /// <param name="b">Blue component value from 0 to 255.</param>
         public Color(int r, int g, int b)
         {
-            _packedValue = 0;
+            _packedValue = 0xFF000000; // A = 255
+
             if (((r | g | b) & 0xFFFFFF00) != 0)
             {
-                R = (byte)MathHelper.Clamp(r, Byte.MinValue, Byte.MaxValue);
-                G = (byte)MathHelper.Clamp(g, Byte.MinValue, Byte.MaxValue);
-                B = (byte)MathHelper.Clamp(b, Byte.MinValue, Byte.MaxValue);
+                var clampedR = (uint)MathHelper.Clamp(r, Byte.MinValue, Byte.MaxValue);
+                var clampedG = (uint)MathHelper.Clamp(g, Byte.MinValue, Byte.MaxValue);
+                var clampedB = (uint)MathHelper.Clamp(b, Byte.MinValue, Byte.MaxValue);
+
+                _packedValue |= (clampedB << 16) | (clampedG << 8) | (clampedR);
             }
             else
             {
-                R = (byte)r;
-                G = (byte)g;
-                B = (byte)b;
+                _packedValue |= ((uint)b << 16) | ((uint)g << 8) | ((uint)r);
             }
-            A = 255;
         }
 
         /// <summary>
@@ -287,20 +282,18 @@ namespace Microsoft.Xna.Framework
         /// <param name="alpha">Alpha component value from 0 to 255.</param>
         public Color(int r, int g, int b, int alpha)
         {
-            _packedValue = 0;
             if (((r | g | b | alpha) & 0xFFFFFF00) != 0)
             {
-                R = (byte)MathHelper.Clamp(r, Byte.MinValue, Byte.MaxValue);
-                G = (byte)MathHelper.Clamp(g, Byte.MinValue, Byte.MaxValue);
-                B = (byte)MathHelper.Clamp(b, Byte.MinValue, Byte.MaxValue);
-                A = (byte)MathHelper.Clamp(alpha, Byte.MinValue, Byte.MaxValue);
+                var clampedR = (uint)MathHelper.Clamp(r, Byte.MinValue, Byte.MaxValue);
+                var clampedG = (uint)MathHelper.Clamp(g, Byte.MinValue, Byte.MaxValue);
+                var clampedB = (uint)MathHelper.Clamp(b, Byte.MinValue, Byte.MaxValue);
+                var clampedA = (uint)MathHelper.Clamp(alpha, Byte.MinValue, Byte.MaxValue);
+
+                _packedValue = (clampedA << 24) | (clampedB << 16) | (clampedG << 8) | (clampedR);
             }
             else
             {
-                R = (byte)r;
-                G = (byte)g;
-                B = (byte)b;
-                A = (byte)alpha;
+                _packedValue = ((uint)alpha << 24) | ((uint)b << 16) | ((uint)g << 8) | ((uint)r);
             }
         }
 
@@ -316,28 +309,7 @@ namespace Microsoft.Xna.Framework
         /// <param name="alpha"></param>
         public Color(byte r, byte g, byte b, byte alpha)
         {
-            _packedValue = 0;
-            R = r;
-            G = g;
-            B = b;
-            A = alpha;
-        }
-
-        /// <summary>
-        /// Constructs an RGBA color from scalars representing red, green, blue and alpha values.
-        /// </summary>
-        /// <param name="r">Red component value from 0.0f to 1.0f.</param>
-        /// <param name="g">Green component value from 0.0f to 1.0f.</param>
-        /// <param name="b">Blue component value from 0.0f to 1.0f.</param>
-        /// <param name="alpha">Alpha component value from 0.0f to 1.0f.</param>
-        public Color(float r, float g, float b, float alpha)
-        {
-            _packedValue = 0;
-			
-            R = (byte)MathHelper.Clamp(r * 255, Byte.MinValue, Byte.MaxValue);
-            G = (byte)MathHelper.Clamp(g * 255, Byte.MinValue, Byte.MaxValue);
-            B = (byte)MathHelper.Clamp(b * 255, Byte.MinValue, Byte.MaxValue);
-            A = (byte)MathHelper.Clamp(alpha * 255, Byte.MinValue, Byte.MaxValue);
+            _packedValue = ((uint)alpha << 24) | ((uint)b << 16) | ((uint)g << 8) | (r);
         }
 
         /// <summary>
@@ -424,10 +396,7 @@ namespace Microsoft.Xna.Framework
         /// <returns><c>true</c> if the instances are equal; <c>false</c> otherwise.</returns>
         public static bool operator ==(Color a, Color b)
         {
-            return (a.A == b.A &&
-                a.R == b.R &&
-                a.G == b.G &&
-                a.B == b.B);
+            return (a._packedValue == b._packedValue);
         }
 	
 	/// <summary>
@@ -438,7 +407,7 @@ namespace Microsoft.Xna.Framework
         /// <returns><c>true</c> if the instances are not equal; <c>false</c> otherwise.</returns>	
         public static bool operator !=(Color a, Color b)
         {
-            return !(a == b);
+            return (a._packedValue != b._packedValue);
         }
 
         /// <summary>
@@ -1884,7 +1853,7 @@ namespace Microsoft.Xna.Framework
         /// <returns>A <see cref="Color"/> which contains premultiplied alpha data.</returns>
         public static Color FromNonPremultiplied(int r, int g, int b, int a)
         {
-            return new Color((byte)(r * a / 255),(byte)(g * a / 255), (byte)(b * a / 255), a);
+            return new Color(r * a / 255, g * a / 255, b * a / 255, a);
         }
 
         #region IEquatable<Color> Members