Fixes #4824. Added the `Color(byte, byte, byte, byte)` constructor and made `Color(uint)` public. Also added some simple unit tests for the constructors and corrected a minor typo.

In the issue, there was some discussion about a separate `Color.FromBytes()` method, but I opted for another constructor overload, since the parameters must be `byte`s in either case. I did not add a `Color(byte, byte, byte)` overload with constant alpha to avoid excessive overloads - this is a more "advanced" overload after all. This can be changed if it makes the API cleaner, of course.

As expected, the new `byte` overload is clearly (40-60%) faster than the `int` version in a quick microbenchmark.