@polsys 2016-08-16
Implemented the short-circuiting of clamping per the discussion. This improved the `int` overload performance a lot (almost halved the time), though the `byte` overload is still slightly (about 20% in my tests) faster. Also added some clamping tests which I had missed.