@MichaelDePiazzi 2016-08-17
Looks good! :thumbsup:
Just a thought while these optimisations are being done - Would it maybe also be worthwhile setting `_packedvalue` in 1 hit, rather than setting each property individually (which performs additional unnecessary operations)? For example:
```c#
public Color(byte r, byte g, byte b, byte alpha)
{
    _packedValue = CalculatePackedValue(r, g, b, alpha);
}

private static uint CalculatePackedValue(byte r, byte g, byte b, byte a)
{
    return r | ((uint)g << 8) | ((uint)b << 16) | ((uint)a << 24);
}
```