@tomspilman 2016-08-17
@MichaelDePiazzi - I suggest an even more direct approach:
```
public Color(byte r, byte g, byte b, byte alpha)
{
    _packedValue = r | ((uint)g << 8) | ((uint)b << 16) | ((uint)alpha<< 24);
}
```
There is little benefit in a shared packing function... it may or may not inline depending on the platform and build settings...  and it isn't like we will be changing our packing format anytime soon.

In the end bugs are reduced by the unit tests in this type... sharing more code here just makes it potentially slower for no real gain.