@MichaelDePiazzi 2016-08-17
@tomspilman While from a clean code point of view this makes me shudder a little, from a performance point of view I have to agree with your points. And I also agree with you about unit tests. Having good test coverage is most beneficial for helping reduce bugs and regressions.

By the way, here are some benchmarks I did to get an idea of how much faster setting the packed value directly is. (Measuring time taken to create 20 million `Color`'s)

- Original `int` constructor = **3174ms**
- New optimised `int` constructor = **1802ms** _(43% less time, 76% faster)_
- New `byte` constructor = **1512ms** _(52% less time, 110% faster)_
- New optimised `int` constructor using direct set = **1203ms** _(62% less time, 164% faster)_
- New `byte` constructor using direct set = **760ms** _(76% less time, 318% faster)_