@tomspilman 2016-08-17
> While from a clean code point of view this makes me shudder a little

I agree...  I hate seeing code duplicated like this very much, but in this sort of type as well as the vector classes performance is first priority.  The only reason I can tolerate it is knowing the unit tests will alert us to any bugs.