@MichaelDePiazzi 2016-08-17
While we're on the subject of optimisations, the `Color` comparison operators could also potentially be improved.

Current code:
```c#
public static bool operator ==(Color a, Color b)
{
    return (a.A == b.A &&
        a.R == b.R &&
        a.G == b.G &&
        a.B == b.B);
}

public static bool operator !=(Color a, Color b)
{
    return !(a == b);
}
```
My suggestion:
```c#
public static bool operator ==(Color a, Color b)
{
    return (a.PackedValue == b.PackedValue);
}

public static bool operator !=(Color a, Color b)
{
    return (a.PackedValue != b.PackedValue);
}
```
Quick benchmark on the equals operator shows it to be ~150% faster.