@tomspilman 2016-08-17
> the  Color  comparison operators could also potentially be improved.

Absolutely!  It is dumb of us to not be taking advantage of direct access to `_packedValue`.  We should prefer that over the `PackedValue` property and certainly the individual component properties.