@MichaelDePiazzi 2016-08-17
> We should prefer that over the PackedValue property

I thought that also... but for some strange reason, it's slower for me when comparing `_packedValue` directly!? I have no idea why either... :confused:

My benchmarks: (Measuring time taken to do 20 million equals comparisons of equal `Color`'s)
- Original (individual property comparison) = **954ms**
- `PackedValue` compare = **447ms**
- `_packedValue` compare = **708ms**

Could it be something to do with the `[CLSCompliant(false)]` tag on the `PackedValue` property?