@polsys 2016-08-17
I completely agree, the optimizations are worth duplicating some code. Working on it right now...