@tomspilman 2016-08-17
>  it's slower for me when comparing `_packedValue` directly!? 

I suggest checking the JIT optimized assembly code directly to see why.