@MichaelDePiazzi 2016-08-17
> I suggest checking the JIT optimized assembly code directly to see why.

Hmmm... seems like the property version is being optimised to do a direct compare:
```
019A65C0  cmp         ecx,edx  
019A65C2  sete        al  
019A65C5  movzx       eax,al  
019A65C8  ret  
```
Whereas the field version is not for some reason:
```
019A65E0  push        ebp  
019A65E1  mov         ebp,esp  
019A65E3  sub         esp,8  
019A65E6  mov         dword ptr [ebp-4],ecx  
019A65E9  mov         dword ptr [ebp-8],edx  
019A65EC  mov         eax,dword ptr [ebp-4]  
019A65EF  cmp         eax,dword ptr [ebp-8]  
019A65F2  sete        al  
019A65F5  movzx       eax,al  
019A65F8  mov         esp,ebp  
019A65FA  pop         ebp  
019A65FB  ret  
```
No idea why...? :confused: