@MichaelDePiazzi 2016-08-17
Update: If I run it in 64-bit mode, it optimises the field version the same as it does for the property version. So for some odd reason, in 32-bit mode the field comparison is not optimised.