@polsys 2016-08-17
After the optimizations, the constructors run about 2-3 times faster. The `float` and vector constructors just offload work to the `int` version, and are still a lot faster. The equality operators also benefited nicely. I did use `_packedValue` for consistency until we have hard data on the JIT behavior - if that is necessary.

I also noted an inconsistency in our `FromNonPremultiplied` behavior. Added unit tests and changed to match XNA.