@KonajuGames 2016-08-17
> The float and vector constructors just offload work to the int version

How does that work? The int version takes a range of 0-255 and the float
version takes 0.0-1.0.