@polsys 2016-08-17
> How does that work? The int version takes a range of 0-255 and the float
version takes 0.0-1.0.

Just deduplicating the old code:

```
public Color(float r, float g, float b)
    : this((int)(r * 255), (int)(g * 255), (int)(b * 255)) 
{
}
```