@tomspilman 2016-08-21
This good to merge?  @KonajuGames @dellis1972 @MichaelDePiazzi @polsys ?